fix no hibernate session errors
By default group members field is a lazy loaded PersitentSet which doesn't work outside transaction/hibernate session. Because transactions are only on GroupDao so outside GroupDao methods no lazy loading is possible. Switch to FetchMode.JOIN to immediately load group members

To test behaviour of hibernate objects out of session scope there was need to test GroupResource not with mock but with database to ensure there hibernate sessions are created